requirements.txt
biopython
matplotlib
tkinter
